collect arith tests
